requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
mysqlclient==2.2.0
Pillow==10.4.0
python-dateutil==2.8.2
openpyxl==3.1.2
sentry-sdk==1.38.0
